+++ src/GPUGenie/inv_table.h
@@ -465,7 +465,7 @@ public:
          *
          */
 	void
-	virtual build(u64 max_length, bool use_load_balance);
+	virtual build(size_t max_length, bool use_load_balance);
 
 
     /*! \fn int get_posting_list_size(int attr_index, int value)
